+++ IdeaProjects/TSP/build.gradle
@@ -29,4 +29,4 @@ javafx {
     version = '11.0.2'
 }
 
-mainClassName  = 'tsp_delaunay.main'
+mainClassName  = 'tsp.delaunay.main'
